move on* functions and init into service
